+++ baybe/core.py
@@ -156,11 +156,14 @@ class BayBE:
         for option, value in self.config.dict().items():
             string += f"   {option}: {value}\n"
 
-        string += "\n\nSearch Space (Experimental Representation):\n"
-        string += f"{self.searchspace.exp_rep}"
+        string += "\n\nSearch Space (Discrete Part, Experimental Representation):\n"
+        string += f"{self.searchspace.discrete.exp_rep}"
 
-        string += "\n\nSearch Space (Computational Representation):\n"
-        string += f"{self.searchspace.comp_rep}"
+        string += "\n\nSearch Space (Discrete Part, Computational Representation):\n"
+        string += f"{self.searchspace.discrete.comp_rep}"
+
+        string += "\n\nSearch Space (Continuous Part, Boundaries):\n"
+        string += f"{self.searchspace.continuous.bounds}"
 
         string += "\n\nMeasurement Space (Experimental Representation):\n"
         string += f"{self.measurements_exp}"
@@ -322,14 +325,14 @@ class BayBE:
                     f"The parameter '{param.name}' has missing values or NaNs in the "
                     f"provided dataframe. Missing parameter values are not supported."
                 )
-            if ("NUM" in param.type) and (data[param.name].dtype.kind not in "iufb"):
+            if (param.is_numeric) and (data[param.name].dtype.kind not in "iufb"):
                 raise TypeError(
                     f"The numerical parameter '{param.name}' has non-numeric entries in"
                     f" the provided dataframe."
                 )
 
         # Update meta data
-        self.searchspace.mark_as_measured(
+        self.searchspace.discrete.mark_as_measured(
             data, self.config.numerical_measurements_must_be_within_tolerance
         )
 
@@ -351,7 +354,7 @@ class BayBE:
         """
         Provides the recommendations for the next batch of experiments.
         """
-        candidates_exp, candidates_comp = self.searchspace.get_candidates(
+        candidates_exp, candidates_comp = self.searchspace.discrete.get_candidates(
             self.config.allow_repeated_recommendations,
             self.config.allow_recommending_already_measured,
         )
@@ -377,7 +380,7 @@ class BayBE:
         #   reflecting the progress of an experiment, whereas the SearchSpace class
         #   would be a stateless representation of the mathematical search space.
         rec = candidates_exp.loc[idxs, :]
-        self.searchspace.metadata.loc[idxs, "was_recommended"] = True
+        self.searchspace.discrete.metadata.loc[idxs, "was_recommended"] = True
 
         # Query user input
         for target in self.targets: