+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Parsing for continuous parameters
+
+### Changed
+- Parameter class hierarchy
+- SearchSpace has now a discrete and continuous subspace
+
 ## [0.2.2] - 2023-01-13
 ### Added
 - SearchSpace class